BuilderFileAccessController load and save now supports * characters
to reiterate, a * in a file is left as a null square, so it can be
assigned a random letter during level generation. Now
FileAccessController is fully compatible with these. (Save is of special
interest, since you have to be extra careful with logic dealing in
nulls, lest you encounter a nullpointerexception)